+++ src/help.c
@@ -97,7 +97,7 @@ void exit_with_usage(char *myname, char *message, char *arg)
 **   Here come the help page routines
 */
 
-static void HelpProcess(char **, int *);
+static void HelpProcess(char **, size_t *);
 static void HelpAbort(void);
 static void HelpRedisplayLine(int, int, int, int);
 static void add_key_to_buf(char *, int);
@@ -198,7 +198,7 @@ void display_help(char *class, struct action *ktabp)
 	helppage();
 }
 
-static void HelpProcess(char **ppbuf, int *plen)
+static void HelpProcess(char **ppbuf, size_t *plen)
 {
 	int done = 0;
 
@@ -409,7 +409,7 @@ static void HelpRedisplayLine(int y, int xs, int xe, int isblank)
 **
 */
 
-static void CopyrightProcess(char **, int *);
+static void CopyrightProcess(char **, size_t *);
 static void CopyrightRedisplayLine(int, int, int, int);
 static void CopyrightAbort(void);
 static void copypage(void);
@@ -462,7 +462,7 @@ screen-devel@gnu.org\n\n\n"
 #endif
     ;
 
-static void CopyrightProcess(char **ppbuf, int *plen)
+static void CopyrightProcess(char **ppbuf, size_t *plen)
 {
 	int done = 0;
 	struct copydata *copydata;
@@ -599,7 +599,7 @@ static void CopyrightRedisplayLine(int y, int xs, int xe, int isblank)
 **
 */
 
-static void BindkeyProcess(char **, int *);
+static void BindkeyProcess(char **, size_t *);
 static void BindkeyAbort(void);
 static void BindkeyRedisplayLine(int, int, int, int);
 static void bindkeypage(void);
@@ -730,7 +730,7 @@ static void bindkeypage()
 	LaySetCursor();
 }
 
-static void BindkeyProcess(char **ppbuf, int *plen)
+static void BindkeyProcess(char **ppbuf, size_t *plen)
 {
 	int done = 0;
 	struct bindkeydata *bindkeydata;
